Check Student Submission

check_sub() is a generic function that check student submission(s) from Moodle Quiz report(s) (i.e. Grades or Responses report). This function can encode & filter student's attempts by state (i.e., "Finished" or "In progress") and started time ("Started on" column). This function also cleans column names of Moodle Quiz report for easier manipulation, extracts student ID from "Email address" column using regular expression, and unites "First name" and "Surname" column into "Name".

Usage

check_sub(
  data,
  extract_id_from = c("Email address", "Institution", "Department", "ID number"),
  id_regex = ".*",
  sep_name = " ",
  state = c("Finished", "In progress"),
  encode = c(1, 0),
  choose_encode = c("max", "min", "all"),
  choose_time = c("first", "last", "all"),
  ...
)

Arguments

data

A data.frame or named list of data.frame of Moodle Quiz report(s) (i.e. either Grades or Responses report).

extract_id_from

(Character) Choose 1 column to extract ID from

id_regex

(Character) A regular expression used to extract ID from column (choose one) "Email address", "Institution", "Department", or "ID number" in the Moodle Quiz report. The default is ".*" meaning all characters. If your student email addresses has numeric IDs in them, try "[:digit:]+" to extract digits from the email. Note: Regular expression syntax is the same as stringr.

sep_name

A character in the new "Name" column that separate original "First name" and "Surname".

state

A character vector to match values in "State" column of the Moodle Quiz report

encode

An encoding numeric vector corresponding to state. For example: by default, in the "State" column, "Finished" values will be encoded as 1, and "In progress" will be encoded as 0.

choose_encode

A character to filter student's attempt by the encodeing.

"max" (default): return rows that have maximum encoding of each students.

"min": return rows that have minimum encoding of each students.

"all": no filter applied, return all rows.

choose_time

A character to filter student's attempt by started time (determined by "Started on" column in Moodle Quiz report). This filter applies after choose_encode has been applied to the data.

"first" (default): return rows that represent first attempt of each students.

"last": return rows that represent last attempt of each students.

"all": no filter applied, return all rows after applying choose_encode

...

argument sep_col of check_sub.list(), which indicates a character separation between names of list and "state" or "encode" columns.

Value

A data.frame, its output content is determined by class of its first argument: data.

If the data is a data.frame; the output is an encoded, filtered, and cleaned data.frame of Moodle Quiz report.

If the data is a named list of data.frame; the output is the same as previously described, but all Moodle Quiz reports are full-joined together by column "Name" and "ID". And, a new column "Total" is computed by the sum of all "...State" columns.

Examples

# Submission of Data Frame
check_sub(grades_ls$Quiz_1,
          id_regex = "[:digit:]+")
#> # A tibble: 26 × 16
#>    Name            ID    Institution Department State Encode Started            
#>    <chr>           <chr> <lgl>       <lgl>      <chr>  <dbl> <dttm>             
#>  1 Jada Roquemore  017   NA          NA         Fini…      1 2021-06-02 08:28:00
#>  2 Ronin Ali       001   NA          NA         Fini…      1 2021-06-02 14:21:00
#>  3 Jerry Hoffpauir 002   NA          NA         Fini…      1 2021-06-02 16:04:00
#>  4 Nathan Babbitt  026   NA          NA         Fini…      1 2021-06-02 00:00:00
#>  5 Rohith Huynh    024   NA          NA         Fini…      1 2021-06-02 14:47:00
#>  6 Joy Ellis       013   NA          NA         Fini…      1 2021-06-02 03:56:00
#>  7 Jeremy Nelson   023   NA          NA         Fini…      1 2021-06-02 08:09:00
#>  8 Christopher Di… 003   NA          NA         Fini…      1 2021-06-02 00:49:00
#>  9 Israel Munoz    018   NA          NA         Fini…      1 2021-06-02 15:36:00
#> 10 Zubaida al-Att… 022   NA          NA         Fini…      1 2021-06-02 23:11:00
#> # … with 16 more rows, and 9 more variables: Grade_10.00 <dbl>, Q1 <dbl>,
#> #   Q2 <dbl>, Q3 <dbl>, Q4 <dbl>, Q5 <dbl>, Q6 <dbl>, Q7 <chr>, Q8 <dbl>

# Submission of List of DF
check_sub(grades_ls,
          id_regex = "[:digit:]+")
#> # A tibble: 26 × 9
#>    Name              ID    Quiz_1_State Quiz_1_Encode Quiz_2_State Quiz_2_Encode
#>    <chr>             <chr> <chr>                <dbl> <chr>                <dbl>
#>  1 Jada Roquemore    017   Finished                 1 Finished                 1
#>  2 Ronin Ali         001   Finished                 1 Finished                 1
#>  3 Jerry Hoffpauir   002   Finished                 1 Finished                 1
#>  4 Nathan Babbitt    026   Finished                 1 Finished                 1
#>  5 Rohith Huynh      024   Finished                 1 Finished                 1
#>  6 Joy Ellis         013   Finished                 1 Finished                 1
#>  7 Jeremy Nelson     023   Finished                 1 Finished                 1
#>  8 Christopher Dill… 003   Finished                 1 Finished                 1
#>  9 Israel Munoz      018   Finished                 1 Finished                 1
#> 10 Zubaida al-Attar  022   Finished                 1 Finished                 1
#> # … with 16 more rows, and 3 more variables: Quiz_3_State <chr>,
#> #   Quiz_3_Encode <dbl>, Total <dbl>