Combine Student's Responses

combine_resp() is a generic function that combine student responses from Moodle Responses report(s) (not Grades report). If the Moodle responses report has response column(s) that contained embedded answers (Cloze), you have an option to split them into different parts (one column for each). Other functionality are similar to check_sub() such as: encode & filter student's attempts by state (i.e., "Finished" or "In progress") and started time ("Started on" column), cleans column names for easy manipulation, extracts student ID from "Email address", and unites "First name" and "Surname" column into "Name".

Usage

combine_resp(
  data,
  extract_id_from = c("Email address", "Institution", "Department", "ID number"),
  id_regex = ".*",
  sep_name = " ",
  state = c("Finished", "In progress"),
  encode = c(1, 0),
  choose_encode = c("max", "min", "all"),
  choose_time = c("first", "last", "all"),
  split_cloze = F,
  part_glue = "_part_",
  sep_col = "_p_"
)

Arguments

data

A data.frame or named list of data.frame of Moodle Responses report(s) (not Grades report)

extract_id_from

(Character) Choose 1 column to extract ID from

id_regex

(Character) A regular expression used to extract ID from column "Email address" in the Moodle Quiz report. The default is ".*" meaning all characters. If your student email addresses has numeric IDs in them, try "[:digit:]+" to extract digits from the email. Note: Regular expression syntax is the same as stringr.

sep_name

A character in the new "Name" column that separate original "First name" and "Surname".

state

A character vector to match values in "State" column of the Moodle Quiz report

encode

An encoding numeric vector corresponding to state. For example: by default, in the "State" column, "Finished" values will be encoded as 1, and "In progress" will be encoded as 0.

choose_encode

A character to filter student's attempt by the encodeing.

"max" (default): return rows that have maximum encoding of each students.

"min": return rows that have minimum encoding of each students.

"all": no filter applied, return all rows.

choose_time

A character to filter student's attempt by started time (determined by "Started on" column in Moodle Quiz report). This filter applies after choose_encode has been applied to the data.

"first" (default): return rows that represent first attempt of each students.

"last": return rows that represent last attempt of each students.

"all": no filter applied, return all rows after applying choose_encode

split_cloze

(Logical) If the data has at least one response column(s) that contained embedded answers (Cloze), split_cloze = TRUE will split them into different parts (one column for each). If the data has no Cloze response columns, split_cloze = TRUE will result in error message. If FALSE (default), no Cloze column will be split.

part_glue

(Character) If split_cloze = TRUE and cloze response columns is presented, part_glue represent a character that glues the names of "Response" column and each "Parts" of Cloze columns together.

sep_col

(Character) If data is a named list of data.frame, sep_col indicate a character separation between names of list and "State" or "Response" columns.

Value

A data.frame, its output content is determined by class of its first argument: data.

If the data is a data.frame; the output is an encoded, filtered, and cleaned data.frame of Moodle Responses report.

If the data is a named list of data.frame; the output is the same as previously described, but all Moodle Quiz reports are full-joined together by column "Name" and "ID". So that, "Response" columns from each data.frame are sit together in a single data.frame. The resulting response columns are in this format (default): "<listNames>-Responses-<number>"

Examples

# Combine Response from 1 Quiz
combine_resp(responses_ls$Quiz_1, id_regex = "[:digit:]+")
#> # A tibble: 26 × 10
#>    Name            ID    Institution Department State Encode Started            
#>    <chr>           <chr> <lgl>       <lgl>      <chr>  <dbl> <dttm>             
#>  1 Lance Xiong     004   NA          NA         Fini…      1 2022-04-19 19:19:00
#>  2 Sydney Halamic… 002   NA          NA         Fini…      1 2022-04-19 16:32:00
#>  3 Itzel Palomino  016   NA          NA         Fini…      1 2022-04-19 12:36:00
#>  4 Hakam el-Farooq 001   NA          NA         Fini…      1 2022-04-19 06:46:00
#>  5 Hamood al-Farha 012   NA          NA         Fini…      1 2022-04-19 03:23:00
#>  6 Dominique Solis 007   NA          NA         Fini…      1 2022-04-19 09:03:00
#>  7 Zhjade Alarid   009   NA          NA         Fini…      1 2022-04-19 02:17:00
#>  8 Jordan Tan      013   NA          NA         Fini…      1 2022-04-19 12:10:00
#>  9 Kevin Vasquez   025   NA          NA         Fini…      1 2022-04-19 03:41:00
#> 10 Angela Carrillo 018   NA          NA         Fini…      1 2022-04-19 01:48:00
#> # … with 16 more rows, and 3 more variables: Grade_2.00 <chr>,
#> #   Response_1 <chr>, Response_2 <chr>

# Combine Response from multiple quizzes into 1 data frame
combine_resp(responses_ls, id_regex = "[:digit:]+")
#> # A tibble: 26 × 9
#>    Name        ID    Quiz_1_State Quiz_1_Response… Quiz_1_Response… Quiz_2_State
#>    <chr>       <chr> <chr>        <chr>            <chr>            <chr>       
#>  1 Lance Xiong 004   Finished     intuitively sin… she replied. “T… Finished    
#>  2 Sydney Hal… 002   Finished     Tarzan’s face w… the means of te… Finished    
#>  3 Itzel Palo… 016   Finished     not to harm you… -                Finished    
#>  4 Hakam el-F… 001   Finished     “A derelict,” w… “I fear,” mused… Finished    
#>  5 Hamood al-… 012   Finished     At last they pa… Finding that it… Finished    
#>  6 Dominique … 007   Finished     the powers of t… though he tried… Finished    
#>  7 Zhjade Ala… 009   Finished     upon the unsusp… and fortune of … Finished    
#>  8 Jordan Tan  013   Finished     “In time to sav… warriors filing… Finished    
#>  9 Kevin Vasq… 025   Finished     A FRIEND.        Thus, making hi… Finished    
#> 10 Angela Car… 018   Finished     very faithful a… “they were kill… Finished    
#> # … with 16 more rows, and 3 more variables: Quiz_2_Response_1 <chr>,
#> #   Quiz_2_Response_2 <chr>, Quiz_2_Response_3 <chr>